requirements.txt
fastapi>=0.100.0
uvicorn>=0.30.0
python-multipart>=0.0.9
openai-whisper>=20231117
